Switch to newly extracted LinkType module.
